+++ R/setup.R
@@ -4,13 +4,13 @@
 #' Users comfortable with reticulate can configure Python manually using `reticulate::use_`.
 #' Note: If users wish to set a custom binary/environment, this must be completed prior to the first call to either `train()` or `complete()`. The same is true if users use the reticulate package directly.
 #' If users wish to switch to a different Python binaries, R must be restarted prior to calling this function.
-#' @param python Character string, path to python binary, or directory of virtualenv, or name of conda environment
+#' @param x Character string, path to python binary, or directory of virtualenv, or name of conda environment
 #' @param type Character string, specifies whether to set a python binary ("auto"), "virtualenv", or "conda"
 #' @param ... Further arguments passed to `reticulate::use_condaenv()`
 #' @keywords setup
 #' @export
 #' @return Boolean indicating whether the custom python environment was activated.
-set_python_env <- function(python, type = "auto", ...) {
+set_python_env <- function(x, type = "auto", ...) {
 
   set_complete <- FALSE
 
@@ -19,17 +19,17 @@ set_python_env <- function(python, type = "auto", ...) {
     return(set_complete)
   } else if (type == "auto") {
 
-    set_py_attempt <- try(reticulate::use_python(python = python, required = TRUE),
+    set_py_attempt <- try(reticulate::use_python(python = x, required = TRUE),
                             silent = TRUE)
 
   } else if (type == "virtualenv") {
 
-    set_py_attempt <- try(reticulate::use_virtualenv(virtualenv = python, required = TRUE),
+    set_py_attempt <- try(reticulate::use_virtualenv(virtualenv = x, required = TRUE),
                           silent = TRUE)
 
   } else if (type == "conda") {
 
-    set_py_attempt <- try(reticulate::use_condaenv(condaenv = python, required = TRUE, ...),
+    set_py_attempt <- try(reticulate::use_condaenv(condaenv = x, required = TRUE, ...),
                           silent = TRUE)
 
   } else {
@@ -37,8 +37,8 @@ set_python_env <- function(python, type = "auto", ...) {
     class(set_py_attempt) <- "try-error"
   }
 
-  if ("try-error" %in% class(set_py_attempt)) {
-    stop("Setting user-specified python environment '",python, "' failed.
+  if (inherits(set_py_attempt,"try-error")) {
+    stop("Setting user-specified python environment '",x, "' failed.
          Please check the specified path/environment and try again.")
   }
 
@@ -70,7 +70,7 @@ python_init <- function() {
         load_stat <- substr(py_config()$version[1],1,1)
       }
 
-      if ("try-error" %in% class(load_stat)) {
+      if (inherits(load_stat, "try-error")) {
 
         stop("Unable to initialise Python and required packages.\n
             Please use set_python_env() to set the Python environment manually, then try again.")
@@ -129,7 +129,7 @@ mid_py_setup <- function() {
   py_pkg_load <- sapply(py_pkgs, function (py_pkg) try(reticulate::import(py_pkg, delay_load = FALSE),
                                                        silent = TRUE))
 
-  missing_pkg <- sapply(py_pkg_load, function (x) ("try-error" %in% class(x)))
+  missing_pkg <- sapply(py_pkg_load, function (x) inherits(x, "try-error"))
   missing_pkg <- py_dep[missing_pkg]
 
   if ("sklearn" %in% missing_pkg) {
@@ -155,7 +155,7 @@ mid_py_setup <- function() {
                                                   python_version = "<3.9"),
                            silent = TRUE)
 
-        if ("try-error" %in% class(pkg_install)) {
+        if (inherits(pkg_install, "try-error")) {
           stop("Unable to install package ", py_pkg, "\n")
         }
       }
@@ -176,7 +176,7 @@ mid_py_setup <- function() {
     if (py_v == "3.9") {
 
       warning("Packages installed but the R session needs to be restarted before proceeding.
-              Please restart R then call set_py_env('your_conda_name', type = 'conda')
+              Please restart R then call set_py_env('your_conda_name', type = 'conda').
               rMIDAS will then be ready to train and impute missing data.")
 
     } else {
@@ -184,7 +184,7 @@ mid_py_setup <- function() {
       py_pkg_load <- sapply(py_pkgs, function (py_pkg) try(reticulate::import(py_pkg, delay_load = FALSE),
                                                            silent = TRUE))
 
-      inst_check <- sum(sapply(py_pkg_load, function (x) ("try-error" %in% class(x))))
+      inst_check <- sum(sapply(py_pkg_load, function (x) inherits(x,"try-error")))
 
       if (inst_check != 0) {
         stop("\nUnable to load required packages after install")